s64 return type for sceKernelTryLockLwMutex

+++ include/usersystemlib_kernel.h
@@ -22,8 +22,8 @@ s32 sceKernelIsCpuIntrEnable(void);
 
 s32 sceKernelGetThreadId(void);
 s32 sceKernelCheckThreadStack(void);
-s32 sceKernelTryLockLwMutex(SceLwMutex *mutex, u32 count);
-s32 sceKernelTryLockLwMutex_600(SceLwMutex *mutex, u32 count);
+s64 sceKernelTryLockLwMutex(SceLwMutex *mutex, u32 count);
+s64 sceKernelTryLockLwMutex_600(SceLwMutex *mutex, u32 count);
 
 void *sceKernelMemcpy(void *dst, const void *src, u32 size);
 void *sceKernelMemset(void *dst, s32 val, u32 size);
